LPD-60291 Adapt the WizardPage component so that it can be reused across other data source connections
