+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/components/salesforce/steps/ConnectSalesforceStep.tsx
@@ -2,21 +2,21 @@ import ClayAlert from '@clayui/alert';
 import ClayForm from '@clayui/form';
 import React from 'react';
 import {Alert} from 'shared/types';
-import {ButtonGroup} from './ButtonGroup';
 import {ConnectSalesforceAuth} from 'settings/components/salesforce/ConnectSalesforceAuth';
 import {DataSourceStatuses} from 'shared/util/constants';
 import {disconnect, fetch} from 'shared/api/data-source';
 import {modalTypes} from 'shared/actions/modals';
 import {Routes, toRoute} from 'shared/util/router';
 import {Text} from '@clayui/core';
-import {updateSearchParams} from '../utis';
-import {useConnectSalesforce} from '../ConnectSalesforceContext';
+import {updateSearchParams} from '../../base-page/utis';
 import {useHistory} from 'react-router-dom';
 import {useQueryParams} from 'shared/hooks/useQueryParams';
+import {useWizardPage} from '../../base-page/WizardPageContext';
+import {WizardPageButtonGroup} from 'settings/components/base-page/WizardPageButtonGroup';
 
 const ConnectSalesforceStep = ({addAlert, close, groupId, onNext, open}) => {
 	const history = useHistory();
-	const {dataSource, setDataSource} = useConnectSalesforce();
+	const {dataSource, setDataSource} = useWizardPage();
 	const {dataSourceId} = useQueryParams();
 
 	if (dataSourceId && !dataSource) {
@@ -72,7 +72,7 @@ const ConnectSalesforceStep = ({addAlert, close, groupId, onNext, open}) => {
 					onSubmit={onNext}
 				/>
 
-				<ButtonGroup
+				<WizardPageButtonGroup
 					nextButtonLabel={Liferay.Language.get('continue')}
 					onCancel={() => {
 						open(modalTypes.CONFIRMATION_MODAL, {
